refactor and bug fix for new repeat step
