+++ src/test/java/com/tinkerpop/gremlin/elastic/elastic/PerformanceTests.java
@@ -5,10 +5,15 @@ import com.tinkerpop.gremlin.LoadGraphWith;
 import com.tinkerpop.gremlin.elastic.ElasticGraphGraphProvider;
 import com.tinkerpop.gremlin.elastic.elasticservice.*;
 import com.tinkerpop.gremlin.elastic.structure.ElasticGraph;
+import com.tinkerpop.gremlin.process.Path;
 import com.tinkerpop.gremlin.process.T;
 import com.tinkerpop.gremlin.process.graph.GraphTraversal;
+import com.tinkerpop.gremlin.process.graph.step.branch.RepeatStep;
+import com.tinkerpop.gremlin.process.graph.step.branch.RepeatTest;
 import com.tinkerpop.gremlin.process.graph.step.filter.WhereStep;
 import com.tinkerpop.gremlin.process.graph.step.filter.WhereTest;
+import com.tinkerpop.gremlin.process.graph.step.sideEffect.CountStep;
+import com.tinkerpop.gremlin.process.graph.step.sideEffect.CountTest;
 import com.tinkerpop.gremlin.structure.*;
 import org.apache.commons.configuration.BaseConfiguration;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
@@ -29,12 +34,12 @@ public class PerformanceTests {
 
 
     @Test
-    @LoadGraphWith(GRATEFUL)
+    @LoadGraphWith(MODERN)
     public void testToPassTests() throws IOException, NoSuchMethodException {
         BaseConfiguration config = new BaseConfiguration();
         config.addProperty(Graph.GRAPH, ElasticGraph.class.getName());
         config.addProperty("elasticsearch.cluster.name", "testgraph");
-        String indexName = "graphtest5";
+        String indexName = "graphtest8";
         config.addProperty("elasticsearch.index.name", indexName.toLowerCase());
         config.addProperty("elasticsearch.local", true);
         config.addProperty("elasticsearch.refresh", true);
@@ -45,7 +50,6 @@ public class PerformanceTests {
         ElasticGraphGraphProvider elasticGraphProvider = new ElasticGraphGraphProvider();
         Method m = this.getClass().getMethod("testToPassTests");
         LoadGraphWith[] loadGraphWiths = m.getAnnotationsByType(LoadGraphWith.class);
-
         elasticGraphProvider.loadGraphData(graph, loadGraphWiths[0], this.getClass(), m.getName());
 
         //GraphTraversal<Vertex, Object> iter = graph.V().has("age").select("name");
@@ -57,13 +61,15 @@ public class PerformanceTests {
         startWatch("graph repeat");
 //        //8 took 127 , 4 took 5.51
         //14465066L
-       // GraphTraversal<Vertex, Long> iter = graph.V().out().count();
+       //GraphTraversal<Vertex, Long> iter = graph.V().out().count();
 
-        GraphTraversal<Vertex, Long> iter = graph.V().repeat(__.out()).times(8).count();
+        //GraphTraversal<Vertex, Long> iter = graph.V().repeat(__.out()).times(3).count();
         //GraphTraversal<Vertex, Object> iter = graph.V().match("a", __.as("a").out().as("b")).select("b").by(T.id);
         //GraphTraversal<Vertex, Element> iter = graph.V().has(T.label, "person").has("name", "marko");
         //GraphTraversal<Vertex, Element> iter = graph.V().both().has(T.label, "software");
-        //GraphTraversal<Vertex, Map<String, Object>> iter = graph.V().has("age").as("a").out().in().has("age").as("b").select().where("a", Compare.eq, "b");
+       //GraphTraversal<Vertex, Map<String, Object>> iter = graph.V().has("age").as("a").out().in().has("age").as("b").select().where("a", Compare.eq, "b");
+        //GraphTraversal<Vertex, Map<String, Object>> iter = graph.V().as("a").out().as("b").select().by("name");
+        GraphTraversal<Vertex, Path> iter = graph.V().emit().times(2).repeat(__.out()).path();
 //        String[] names = new String[2];
 //        names[0]="marko";
 //        names[1]="josh";
